Separated onboarding into two pages
